refactor: standardize DI naming to match AI_CONTEXT.md convention
- Rename all DI entry files to lib/core/di/di.dart
- Rename all generated files to di.module.dart
- Standardize all locator names to getIt (was ccSdkLocator, appConfigLocator, getItData)
- Flatten data module DI directory (remove inject/ subdirectory)
- Update all imports and exports across modules, libraries, and main app
- Update documentation (README, feature_template, module READMEs)
