+++ README.md
@@ -320,7 +320,7 @@ Template
     │    │    │    ├── di
     │    │    │    │    ├── module
     │    │    │    │    │      └── ...
-    │    │    │    │    └── injection.dart
+    │    │    │    │    └── di.dart
     │    │    │    ├── network
     │    │    │    │    └── ...
     │    │    │    ├── page
