+++ lib/sample_code/getx_simple_page/way_1/getx/get_view_controller.dart
@@ -39,7 +39,7 @@ class GetViewController extends CcGetController {
   final RxList<ResSampleCodeFakeModel> sampleCodeFakeList =
       <ResSampleCodeFakeModel>[].obs; // init: data observable
 
-  // SampleCodeFakeApiImpl repositories = getItData<SampleCodeFakeApiImpl>();
+  // SampleCodeFakeApiImpl repositories = getIt<SampleCodeFakeApiImpl>();
 
   @override
   void onInit() {
