+++ libraries/cc_sdk/lib/core/di/di.dart
@@ -1,7 +1,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
-final GetIt ccSdkLocator = GetIt.instance;
+final GetIt getIt = GetIt.instance;
 
 @InjectableInit.microPackage()
 void initMicroPackage() {}
