+++ modules/app_config/lib/core/di/di.dart
@@ -1,7 +1,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
-final GetIt appConfigLocator = GetIt.instance;
+final GetIt getIt = GetIt.instance;
 
 /// Configures dependency injection for the app_config module.
 /// Uses the Micro-Package pattern for injectable.
