+++ modules/app_config/lib/core/di/export_di_app_config.dart
@@ -1,5 +1,5 @@
 /// This library provides access to the AppConfig module's dependency injection system.
 library app_config_di;
 
-export 'di_app_config.dart';
+export 'di.dart';
 export 'module/dependencies.dart';
