+++ modules/app_config/lib/core/extension/app_track_log_extension.dart
@@ -1,18 +1,18 @@
 import 'package:cc_sdk/core/extensions/export_extensions.dart';
 
 import '../../data/datasource/local/box/app_track_log/cc_app_track_log.dart';
-import '../di/di_app_config.dart';
+import '../di/di.dart';
 
 extension AppTrackLogExtension on CcAppTrackLog {
   List<String>? initMsgIfNull() {
-    var list = appConfigLocator<CcAppTrackLog>().msg;
+    var list = getIt<CcAppTrackLog>().msg;
     try {
       list ??= [];
     } catch (e) {
       'AppTrackLogExtension : e'.Log();
     }
 
-    appConfigLocator<CcAppTrackLog>().msg = list;
+    getIt<CcAppTrackLog>().msg = list;
 
     return list;
   }
