+++ modules/data/lib/core/di/di_export.dart
@@ -1,4 +1,4 @@
 library data_di;
 
-export 'inject/data_inject.dart';
+export 'di.dart';
 export 'module/data_module.dart';
